Fix / remove Head component
